+++ lib/_module.nix
@@ -2,6 +2,19 @@
 {lib, ...}: let
   map = import ./map.nix lib;
   build = import ./build.nix lib;
+
+  # Flake configurations type
+  types.configuration = lib.types.submodule {
+    options = {
+      module = lib.mkOption {
+        type = lib.types.deferredModule;
+      };
+      system = lib.mkOption {
+        type = lib.types.str;
+        default = "x86_64-linux";
+      };
+    };
+  };
 in {
   options.util = lib.mkOption {
     type = lib.types.anything;
@@ -10,6 +23,6 @@ in {
   };
 
   config.util = {
-    inherit map build;
+    inherit map build types;
   };
 }