+++ modules/gui/theme.nix
@@ -13,29 +13,23 @@ in {
       pkgs,
       ...
     }: let
-      inherit
-        (lib)
-        attrNames
-        mkOption
-        optionals
-        types
-        ;
       cfg = config.gui;
-      inherit (config.gui) enable;
+      stl = config.stylix;
+      inherit (cfg) enable;
     in {
       imports = [inputs.stylix.nixosModules.stylix];
 
       options.gui = {
         enable = lib.mkEnableOption "Graphical Desktop Session";
         fancy = lib.mkEnableOption "Enable Fancy GUI Effects";
-        display = mkOption {
+        display = lib.mkOption {
           description = "Main GUI Display";
-          type = types.str;
+          type = lib.types.str;
           default = "eDP-1";
           example = "HDMI-A-1";
         };
-        wallpaper = mkOption {
-          type = types.enum (attrNames files.wallpapers);
+        wallpaper = lib.mkOption {
+          type = lib.types.enum (lib.attrNames files.wallpapers);
           default = "Beauty";
           apply = image: files.wallpapers."${image}";
         };
@@ -55,7 +49,7 @@ in {
 
           icons = {
             enable = true;
-            package = pkgs.papirus-icon-theme;
+            package = lib.mkDefault pkgs.papirus-icon-theme;
             light = "Papirus-Dark";
             dark = "Papirus-Dark";
           };
@@ -78,8 +72,13 @@ in {
           };
         };
 
-        home-manager.sharedModules = optionals (!enable) [
+        home-manager.sharedModules = lib.optionals (!enable) [
           config.stylix.homeManagerIntegration.module
+        ];
+
+        environment.systemPackages = lib.optionals enable [
+          stl.icons.package
+          stl.cursor.package
         ];
 
         programs.gnupg.agent.pinentryPackage = pkgs.pinentry-gtk2;