+++ modules/hardware/mouse.nix
@@ -1,0 +1,14 @@
+## Mouse Firmware ##
+_: {
+  flake.modules = {
+    nixos.mouse = _: {
+      hardware.logitech.wireless = {
+        enable = true;
+        enableGraphical = true;
+      };
+    };
+    homeManager.mouse = _: {
+      home.persist.directories = [".config/solaar"];
+    };
+  };
+}