+++ lib/_module.nix
@@ -1,21 +1,5 @@
 ## Custom Library Functions ##
-{lib, ...}: let
-  map = import ./map.nix lib;
-  build = import ./build.nix lib;
-
-  # Flake configurations type
-  types.configuration = lib.types.submodule {
-    options = {
-      module = lib.mkOption {
-        type = lib.types.deferredModule;
-      };
-      system = lib.mkOption {
-        type = lib.types.str;
-        default = "x86_64-linux";
-      };
-    };
-  };
-in {
+{lib, ...}: {
   options.util = lib.mkOption {
     type = lib.types.anything;
     readOnly = true;
@@ -23,6 +7,8 @@ in {
   };
 
   config.util = {
-    inherit map build types;
+    map = import ./map.nix lib;
+    build = import ./build.nix lib;
+    types = import ./types.nix lib;
   };
 }